add repository arg type

+++ src/main/java/com/contract/domain/ContractRepository.java
@@ -2,6 +2,6 @@ package com.contract.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Integer>{
+public interface ContractRepository extends JpaRepository<Contract, Long>{
     
 }
